Added TestBaseWKBSerdeJava for testing using the WKB serde Moved WBBGeometrySerializerTest from the scala to java to eliminate spark session collisions
